move space packet dir
